requirements.txt
flask
flask-wtf
email_validator
wtforms
requests
requests-ratelimiter
requests-oauthlib
python-dotenv
schedule
shortuuid
gunicorn